6/23 cleaner

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/data/Product.java
@@ -20,7 +20,7 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private int amount;
+    private Long amount;
     private String unit;
     private double price;
